requirements.txt
streamlit==1.24.0
tensorflow==2.14.0
numpy==1.26.0
Pillow==10.0.0
plotly==5.17.0
matplotlib==3.8.0
seaborn==0.13.2
scikit-learn==1.3.0
gTTS==2.3.2
pyttsx3==2.90
opencv-python==4.8.0.76
